requirements.txt
# Erebus Python dependencies
pywebview>=4.4.1
# tomllib is stdlib in Python 3.11+; install tomli as backport for older versions:
# tomli>=2.0.0

# ── Font dependencies (non-Python, install via your OS package manager) ────────
# Erebus defaults to JetBrains Mono. On Arch: pacman -S ttf-jetbrains-mono
# Fallbacks: Fira Code (ttf-fira-code), Cascadia Code (ttf-cascadia-code)
# Without any of these, the system monospace font will be used instead.
# A web font fallback is bundled in the UI via Google Fonts CDN (requires internet).

# ── Optional runtime tools ─────────────────────────────────────────────────────
# ripgrep  — faster find-in-files search (falls back to pure Python if absent)
# git      — git status badges and diff in the explorer
